Fix incorrect type assertion in He forward model

+++ src/he_forward.jl
@@ -34,7 +34,7 @@ Calculate radioactive ingrowth and diffusion of He4 using modified eqns
 - `eta_q`: product η*q; η - fission-track etching efficiency, q - scaling factor
 
 """
-function rdaam_forward_diffusion(alpha::Float64,c0,c1,c2,c3,rmr0,eta_q,L_dist,psi,omega,Etrap,
+function rdaam_forward_diffusion(alpha,c0,c1,c2,c3,rmr0,eta_q,L_dist,psi,omega,Etrap,
   R,Ea,logD0_a2,n_iter,U238,U238_V,U235,U235_V,Th232,Th232_V,L,
   tT...)
   tTcopy = collect(tT)
